Abstract IC object function generation
