+++ src/compiler/typescript_to_rust/generators/ic_object/functions/caller.ts
@@ -1,7 +1,7 @@
 import { Rust } from '../../../../../types';
 
 export function generateIcObjectFunctionCaller(): Rust {
-    return `
+    return /* rust */ `
         fn azle_ic_caller(
             _this: &boa_engine::JsValue,
             _aargs: &[boa_engine::JsValue],
