+++ src/compiler/typescript_to_rust/generators/ic_object/functions/canister_balance.ts
@@ -1,7 +1,7 @@
 import { Rust } from '../../../../../types';
 
 export function generateIcObjectFunctionCanisterBalance(): Rust {
-    return `
+    return /* rust */ `
         fn _azle_ic_canister_balance(
             _this: &boa_engine::JsValue,
             _aargs: &[boa_engine::JsValue],
