+++ src/compiler/typescript_to_rust/generators/ic_object/functions/id.ts
@@ -1,7 +1,7 @@
 import { Rust } from '../../../../../types';
 
 export function generateIcObjectFunctionId(): Rust {
-    return `
+    return /* rust */ `
         fn azle_ic_id(
             _this: &boa_engine::JsValue,
             _aargs: &[boa_engine::JsValue],
