+++ src/compiler/typescript_to_rust/generators/ic_object/functions/index.ts
@@ -1,0 +1,35 @@
+import { Rust } from '../../../../../types';
+
+import { generateIcObjectFunctionAcceptMessage } from './accept_message';
+import { generateIcObjectFunctionCaller } from './caller';
+import { generateIcObjectFunctionCanisterBalance } from './canister_balance';
+import { generateIcObjectFunctionCanisterBalance128 } from './canister_balance128';
+import { generateIcObjectFunctionId } from './id';
+import { generateIcObjectFunctionMethodName } from './method_name';
+import { generateIcObjectFunctionPrint } from './print';
+import { generateIcObjectFunctionTime } from './time';
+import { generateIcObjectFunctionTrap } from './trap';
+
+export function generateIcObjectFunctions(): Rust {
+    const icObjectFunctionAcceptMessage: Rust = generateIcObjectFunctionAcceptMessage();
+    const icObjectFunctionCaller: Rust = generateIcObjectFunctionCaller();
+    const icObjectFunctionCanisterBalance: Rust = generateIcObjectFunctionCanisterBalance();
+    const icObjectFunctionCanisterBalance128: Rust = generateIcObjectFunctionCanisterBalance128();
+    const icObjectFunctionId: Rust = generateIcObjectFunctionId();
+    const icObjectFunctionMethodName: Rust = generateIcObjectFunctionMethodName();
+    const icObjectFunctionPrint: Rust = generateIcObjectFunctionPrint();
+    const icObjectFunctionTime: Rust = generateIcObjectFunctionTime();
+    const icObjectFunctionTrap: Rust = generateIcObjectFunctionTrap();
+
+    return `
+        ${icObjectFunctionAcceptMessage}
+        ${icObjectFunctionCaller}
+        ${icObjectFunctionCanisterBalance}
+        ${icObjectFunctionCanisterBalance128}
+        ${icObjectFunctionId}
+        ${icObjectFunctionMethodName}
+        ${icObjectFunctionPrint}
+        ${icObjectFunctionTime}
+        ${icObjectFunctionTrap}
+  `;
+}
