+++ src/compiler/typescript_to_rust/generators/ic_object/functions/print.ts
@@ -1,14 +1,14 @@
 import { Rust } from '../../../../../types';
 
 export function generateIcObjectFunctionPrint(): Rust {
-    return `
+    return /* rust */ `
         fn azle_ic_print(
             _this: &boa_engine::JsValue,
             _aargs: &[boa_engine::JsValue],
             _context: &mut boa_engine::Context
         ) -> boa_engine::JsResult<boa_engine::JsValue> {
             ic_cdk::println!("{:#?}", _aargs);
-        
+
             return Ok(boa_engine::JsValue::Undefined);
         }
     `;
